refactor: pages/ 로 폴더구조 개선
